pp: remove sharded_client_cache wrapper
The wrapper is obsolete now since the auth_ctx_server calls the sharded
client cache directly.
