+++ src/v/pandaproxy/CMakeLists.txt
@@ -6,7 +6,6 @@ v_cc_library(
     probe.cc
     server.cc
     kafka_client_cache.cc
-    sharded_client_cache.cc
     config_utils.cc
   DEPS
     v::pandaproxy_parsing
