+++ src/v/pandaproxy/test/CMakeLists.txt
@@ -13,7 +13,6 @@ rp_test(
   BINARY_NAME pandaproxy_single_thread
   SOURCES
     one_shot.cc
-    sharded_client_cache.cc
     kafka_client_cache.cc
   DEFINITIONS BOOST_TEST_DYN_LINK
   LIBRARIES v::seastar_testing_main v::rphashing v::http v::pandaproxy_common v::kafka_protocol
